+++ sessions/backend-create/demo-end/pages/api/jokes/index.js
@@ -1,23 +1,23 @@
-import dbConnect from "../../../db/connect";
-import Joke from "../../../db/models/Joke";
+import dbConnect from "@/db/connect";
+import Joke from "@/db/models/Joke";
 
 export default async function handler(request, response) {
   await dbConnect();
 
   if (request.method === "GET") {
     const jokes = await Joke.find();
-    return response.status(200).json(jokes);
+
+    response.status(200).json(jokes);
+    return;
   }
 
   if (request.method === "POST") {
-    try {
-      const jokeData = request.body;
-      await Joke.create(jokeData);
+    const jokeData = request.body;
+    await Joke.create(jokeData);
 
-      response.status(201).json({ status: "Joke created" });
-    } catch (error) {
-      console.log(error);
-      response.status(400).json({ error: error.message });
-    }
+    response.status(201).json({ status: "Joke created" });
+    return;
   }
+
+  response.status(405).json({ status: "Method not allowed." });
 }
