+++ sessions/backend-create/demo-start/pages/api/jokes/index.js
@@ -1,11 +1,14 @@
-import dbConnect from "../../../db/connect";
-import Joke from "../../../db/models/Joke";
+import dbConnect from "@/db/connect";
+import Joke from "@/db/models/Joke";
 
 export default async function handler(request, response) {
   await dbConnect();
 
   if (request.method === "GET") {
     const jokes = await Joke.find();
-    return response.status(200).json(jokes);
+    response.status(200).json(jokes);
+    return;
   }
+
+  response.status(405).json({ status: "Method not allowed." });
 }
