--- sessions/backend-create/products/components/Product/Product.styled.js
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const ProductCard = styled.article`
-  padding: 0.5rem 1rem;
-  box-shadow: 0px 1px 5px -2px var(--color-granite);
-`;
