--- sessions/backend-create/products/components/ProductForm/ProductForm.styled.js
@@ -1,18 +1,0 @@
-import styled from "styled-components";
-
-export const StyledForm = styled.form`
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-`;
-
-export const StyledHeading = styled.h2`
-  text-align: center;
-  color: var(--color-nemo);
-`;
-
-export const StyledLabel = styled.label`
-  display: flex;
-  justify-content: space-between;
-  gap: 5px;
-`;
