+++ sessions/backend-create/products/pages/[id].js
@@ -1,5 +1,35 @@
-import Product from "../components/product";
+import useSWR from "swr";
+import styled from "styled-components";
+import { useRouter } from "next/router";
+import StyledLink from "@/components/Link";
 
-export default function ProductDetailsPage() {
-  return <Product />;
+export default function Product() {
+  const router = useRouter();
+  const { id } = router.query;
+
+  const { data, isLoading } = useSWR(`/api/products/${id}`);
+
+  if (isLoading) {
+    return <h1>Loading...</h1>;
+  }
+
+  if (!data) {
+    return;
+  }
+
+  return (
+    <ProductCard>
+      <h2>{data.name}</h2>
+      <p>Description: {data.description}</p>
+      <p>
+        Price: {data.price} {data.currency}
+      </p>
+      <StyledLink href="/">Back to all</StyledLink>
+    </ProductCard>
+  );
 }
+
+export const ProductCard = styled.article`
+  padding: 0.5rem 1rem;
+  box-shadow: 0px 1px 5px -2px var(--color-granite);
+`;
