+++ sessions/backend-create/products/pages/api/products/[id].js
@@ -1,5 +1,5 @@
-import dbConnect from "../../../db/connect";
-import Product from "../../../db/models/Product";
+import dbConnect from "@/db/connect";
+import Product from "@/db/models/Product";
 
 export default async function handler(request, response) {
   await dbConnect();
@@ -9,9 +9,13 @@ export default async function handler(request, response) {
     const product = await Product.findById(id).populate("reviews");
 
     if (!product) {
-      return response.status(404).json({ status: "Not Found" });
+      response.status(404).json({ status: "Not Found" });
+      return;
     }
 
     response.status(200).json(product);
+    return;
   }
+
+  response.status(405).json({ status: "Method not allowed." });
 }
