+++ sessions/backend-update-and-delete/demo-start/pages/api/jokes/[id].js
@@ -1,5 +1,5 @@
-import dbConnect from "../../../db/connect";
-import Joke from "../../../db/models/Joke";
+import dbConnect from "@/db/connect";
+import Joke from "@/db/models/Joke";
 
 export default async function handler(request, response) {
   await dbConnect();
