+++ sessions/backend-update-and-delete/products/pages/api/products/[id].js
@@ -1,5 +1,5 @@
-import dbConnect from "../../../db/connect";
-import Product from "../../../db/models/Product";
+import dbConnect from "@/db/connect";
+import Product from "@/db/models/Product";
 
 export default async function handler(request, response) {
   await dbConnect();
